requirements.txt
alabaster==0.7.10
argon2-cffi==18.1.0
attrs==17.4.0
awesome-slugify==1.6.5
Babel==2.5.3
boto3==1.4.7
botocore==1.7.48
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
Collectfast==0.5.2
coverage==4.5.1
decorator==4.2.1
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.35.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==1.9.9
django-model-utils==3.1.1
django-redis==4.8.0
django-storages==1.6.5
django-tenant-schemas==1.9.0
django-test-plus==1.0.22
docutils==0.14
factory-boy==2.10.0
Faker==0.8.11
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
imagesize==1.0.0
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
markdown2==2.3.5
MarkupSafe==1.0
oauthlib==2.0.6
packaging==16.8
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.2
pycparser==2.18
Pygments==2.2.0
pymongo==3.6.1
pyparsing==2.2.0
pytest==3.4.1
pytest-django==3.1.2
pytest-sugar==0.9.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
regex==2018.2.21
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.0
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.1
traitlets==4.3.2
typing==3.6.4
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
